Move models to config
